Chore: Move specification queries to inside the specification component
