+++ react/components/SpecificationsTranslation/graphql/downloadFieldTranslations.gql
