+++ react/components/SpecificationsTranslation/graphql/fieldTranslations.gql
