+++ react/components/SpecificationsTranslation/graphql/fieldTranslationsRequests.gql
