+++ react/components/SpecificationsTranslation/graphql/fieldTranslationsUploadRequests.gql
