+++ react/components/SpecificationsTranslation/graphql/getSpecification.gql
