+++ react/components/SpecificationsTranslation/graphql/translateSpecification.gql
