+++ react/components/SpecificationsTranslation/graphql/uploadFieldTranslationsImport.gql
