--- react/graphql/getSpecificationsValues.gql
@@ -1,6 +1,0 @@
-query GetSpecificationsValues($fieldId: ID) {
-  fieldValues(fieldId: $fieldId) @context(provider: "vtex.catalog-graphql") {
-    fieldValueId
-    value
-  }
-}
