--- react/graphql/translateSpecificationValues.gql
@@ -1,7 +1,0 @@
-mutation translateSpecificationValues(
-  $args: FieldValueInputTranslation
-  $locale: Locale
-) {
-  translateFieldValues(fieldValues: $args, locale: $locale)
-    @context(provider: "vtex.catalog-graphql")
-}
